Only creating gamepad data when gamepad information is available (#539)
* gamepad fix - don't assume everything is a gamepad

* Rework based on better understanding

* auto -> auto&

+++ Dependencies/xr/Include/XR.h
@@ -201,7 +201,7 @@ namespace xr
                     {
                         bool Pressed{ false };
                         bool Touched{ false };
-                        double Value{0};
+                        float Value{0};
                     };
 
                     std::array<float, DEFAULT_CONTROLLER_AXES> Axes;
